Get images from wikipedia

+++ main.py
@@ -1,6 +1,7 @@
 import os
+import random
 import discord
-import urllib.request
+import urllib.request, urllib.parse
 import json
 import garmets_data
 
@@ -33,28 +34,44 @@ if not weather_key:
 
 client = discord.Client()
 
-cache_file_index = 0
-
-def get_weather():
-    # 43.548899,-79.6650758
-    response = urllib.request.urlopen(
-        "https://api.openweathermap.org/data/2.5/weather?lat=43.548899&lon=-79.6650758&appid=" + weather_key)
+def get_JSON(url):
+    response = urllib.request.urlopen(url)
     response_data = response.read()
     return json.loads(response_data)
 
 
-async def send_message_with_unsplash_image(text, query, channel):
+def get_weather():
+    # 43.548899,-79.6650758
+    return get_JSON("https://api.openweathermap.org/data/2.5/weather?lat=43.548899&lon=-79.6650758&appid=" + weather_key)
+
+
+def get_wikimedia_image(query):
+    response_data = get_JSON(
+        "https://en.wikipedia.org/w/api.php?action=query&format=json&prop=images&titles=" +
+        query
+    )
+
+    pages = response_data['query']['pages']
+    images = list(pages.items())[0][1]['images']
+    rand_image = random.choice(images)
+
+    image_data = get_JSON(f"https://en.wikipedia.org/w/api.php?action=query&format=json&prop=imageinfo&titles={urllib.parse.quote(rand_image['title'])}&iiprop=url&iiurlwidth=300")
+    image_pages = image_data['query']['pages']
+    url = list(image_pages.items())[0][1]['imageinfo'][0]['thumburl']
+
+    return url
+
+
+async def send_message_with_image(text, image_url, channel):
     global cache_file_index
 
-    image_req = urllib.request.urlopen(
-        "https://source.unsplash.com/random/280x280/?" + query)
+    cache_file_path = f"./cache/{urllib.parse.quote(image_url, safe='')}"
 
-    cache_file_index += 1
-    _local_cache_file_index = cache_file_index
-    cache_file_path = f"./cache/image{_local_cache_file_index}.jpg"
+    if not os.path.exists(cache_file_path):
+        image_req = urllib.request.urlopen(image_url)
+        with open(cache_file_path, 'wb') as file:
+            file.write(image_req.read())
 
-    with open(cache_file_path, 'wb') as file:
-        file.write(image_req.read())
     with open(cache_file_path, 'rb') as file:
         discord_file = discord.File(file)
         await channel.send(text, file=discord_file)
@@ -83,16 +100,16 @@ async def on_message(message):
     if message.author == client.user:
         return
 
-    if message.content.startswith('$hello'):
-        # await message.channel.send('Hello!')
-        await message.channel.send(get_weather())
-    elif message.content.startswith('$unsplash'):
+    text_clean = message.content.lower()
+
+    if text_clean.startswith('what should i wear'):
         outfit = get_outfit_based_on_weather()
-        await send_message_with_unsplash_image(
+        await send_message_with_image(
             text=outfit[0], 
-            query=outfit[0],
+            image_url=get_wikimedia_image(outfit[0]),
             channel=message.channel
         )
+
 
 client.run(bot_key)
